Separate Handler and Storer by interfaces
Also minor changes into Makefile and logging
